Use graceful-fs and remove lib directory
https://github.com/isaacs/node-graceful-fs
