+++ package.json
@@ -10,7 +10,7 @@
     "test": "test"
   },
   "scripts": {
-    "test": "jshint lib/*.js test/*.js index.js && mocha"
+    "test": "jshint test/*.js index.js && mocha"
   },
   "repository": "git@github.com:jgable/cache-swap.git",
   "keywords": [
@@ -22,6 +22,7 @@
   "author": "Jacob Gable <jacob.gable@gmail.com>",
   "license": "MIT",
   "dependencies": {
+    "graceful-fs": "^3.0.5",
     "mkdirp": "^0.5.0",
     "object-assign": "^2.0.0",
     "rimraf": "^2.2.8"
